Wow So Much Code
So Doge .. Wow

+++ app/src/main/java/com/capsulecorp/project_trackerino/myMarker.java
@@ -7,13 +7,13 @@ import org.osmdroid.views.MapView;
  * Created by Dennis on 02.07.2015.
  */
 public class myMarker extends Marker {
-    int id;
-    public myMarker(MapView mapView, int id) {
+    long id;
+    public myMarker(MapView mapView, long id) {
         super(mapView);
         this.id = id;
     }
 
-    public int getId(){
+    public long getId(){
         return id;
     }
 }
